+++ crates/ruma-macros/src/lib.rs
@@ -15,8 +15,8 @@ use proc_macro::TokenStream;
 use proc_macro2 as pm2;
 use quote::quote;
 use ruma_identifiers_validation::{
-    device_key_id, event_id, key_id, mxc_uri, room_alias_id, room_id, room_version_id, server_name,
-    user_id,
+    base64_public_key, event_id, mxc_uri, room_alias_id, room_id, room_version_id, server_name,
+    server_signing_key_version, user_id,
 };
 use syn::{parse_macro_input, DeriveInput, ItemEnum, ItemStruct};
 
@@ -146,19 +146,6 @@ pub fn derive_id_zst(input: TokenStream) -> TokenStream {
     expand_id_zst(input).unwrap_or_else(syn::Error::into_compile_error).into()
 }
 
-/// Compile-time checked `DeviceKeyId` construction.
-#[proc_macro]
-pub fn device_key_id(input: TokenStream) -> TokenStream {
-    let IdentifierInput { dollar_crate, id } = parse_macro_input!(input as IdentifierInput);
-    assert!(device_key_id::validate(&id.value()).is_ok(), "Invalid device key id");
-
-    let output = quote! {
-        <&#dollar_crate::DeviceKeyId as ::std::convert::TryFrom<&str>>::try_from(#id).unwrap()
-    };
-
-    output.into()
-}
-
 /// Compile-time checked `EventId` construction.
 #[proc_macro]
 pub fn event_id(input: TokenStream) -> TokenStream {
@@ -211,14 +198,17 @@ pub fn room_version_id(input: TokenStream) -> TokenStream {
     output.into()
 }
 
-/// Compile-time checked `ServerSigningKeyId` construction.
+/// Compile-time checked `ServerSigningKeyVersion` construction.
 #[proc_macro]
-pub fn server_signing_key_id(input: TokenStream) -> TokenStream {
+pub fn server_signing_key_version(input: TokenStream) -> TokenStream {
     let IdentifierInput { dollar_crate, id } = parse_macro_input!(input as IdentifierInput);
-    assert!(key_id::validate(&id.value()).is_ok(), "Invalid server_signing_key_id");
+    assert!(
+        server_signing_key_version::validate(&id.value()).is_ok(),
+        "Invalid server_signing_key_version"
+    );
 
     let output = quote! {
-        <&#dollar_crate::ServerSigningKeyId as ::std::convert::TryFrom<&str>>::try_from(#id).unwrap()
+        <&#dollar_crate::ServerSigningKeyVersion as ::std::convert::TryFrom<&str>>::try_from(#id).unwrap()
     };
 
     output.into()
@@ -258,6 +248,19 @@ pub fn user_id(input: TokenStream) -> TokenStream {
 
     let output = quote! {
         <&#dollar_crate::UserId as ::std::convert::TryFrom<&str>>::try_from(#id).unwrap()
+    };
+
+    output.into()
+}
+
+/// Compile-time checked `Base64PublicKey` construction.
+#[proc_macro]
+pub fn base64_public_key(input: TokenStream) -> TokenStream {
+    let IdentifierInput { dollar_crate, id } = parse_macro_input!(input as IdentifierInput);
+    assert!(base64_public_key::validate(&id.value()).is_ok(), "Invalid base64 public key");
+
+    let output = quote! {
+        <&#dollar_crate::DeviceKeyId as ::std::convert::TryFrom<&str>>::try_from(#id).unwrap()
     };
 
     output.into()